Move assessment models/dtos to a separate folder for organizational purposes
