+++ NursingEducationalBackend/NursingEducationalBackend/Models/Assessments/AcuteProgress.cs
@@ -1,9 +1,10 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace NursingEducationalBackend.DTOs
+namespace NursingEducationalBackend.Models.Assessments
 {
-    public class PatientAcuteProgressDTO
+    public class AcuteProgress
     {
+        public int AcuteProgressId { get; set; }
 
         [Range(typeof(DateTime), "1900-01-01 00:00:00", "3000-12-31 00:00:00")]
         public DateTime Timestamp { get; set; }
