+++ NursingEducationalBackend/NursingEducationalBackend/Models/Assessments/Behaviour.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace NursingEducationalBackend.Models.Assessments;
+
+public partial class Behaviour
+{
+    public int BehaviourId { get; set; }
+
+    public string Report { get; set; } = null!;
+}
