+++ NursingEducationalBackend/NursingEducationalBackend/Models/Assessments/ProgressNote.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace NursingEducationalBackend.Models.Assessments;
+
+public partial class ProgressNote
+{
+    public int ProgressNoteId { get; set; }
+
+    [Range(typeof(DateTime), "1900-01-01 00:00:00", "3000-12-31 00:00:00")]
+    public DateTime Timestamp { get; set; }
+
+    public string Note { get; set; } = null!;
+}
