--- NursingEducationalBackend/NursingEducationalBackend/Models/Behaviour.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace NursingEducationalBackend.Models;
-
-public partial class Behaviour
-{
-    public int BehaviourId { get; set; }
-
-    public string Report { get; set; } = null!;
-}
