+++ NursingEducationalBackend/NursingEducationalBackend/Models/NursingDbContext.cs
@@ -5,6 +5,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.AspNetCore.Identity;
+using NursingEducationalBackend.Models.Assessments;
 
 namespace NursingEducationalBackend.Models;
